requirements.txt
colorlog
ecdsa
oslo.i18n==1.5.0
stevedore==1.5.0
oslo.serialization==1.4.0
debtcollector==0.3.0
cliff==1.4.3
fabric>=1.8.2
ipaddr
jinja2
jsondate
PyMySQL==0.6.7
oslo.config==1.15.0
oslo.utils==1.9.0
pika==0.9.14
progressbar==2.3
pycrypto
python-cinderclient==1.1.0
python-glanceclient==0.12.0
python-keystoneclient==1.6.0
python-neutronclient==2.3.4
python-novaclient==2.20.0
python-swiftclient==2.3.1
pyyaml
pywbem
redis
sqlalchemy
marshmallow==2.4.2
